feat : 버튼 클릭시 다음 스텝으로 이동하는 기능 구현
