+++ src/constants/category.js
@@ -1,0 +1,31 @@
+const catetoryCode = [
+  {
+    value: 'all',
+    label: '전체',
+  },
+  {
+    value: 'c1',
+    label: '관광지',
+  },
+  {
+    value: 'c2',
+    label: '쇼핑',
+  },
+  {
+    value: 'c3',
+    label: '숙박',
+  },
+  {
+    value: 'c4',
+    label: '음식',
+  },
+  {
+    value: 'c5',
+    label: '축제/행사',
+  },
+  {
+    value: 'c6',
+    label: '테마여행',
+  },
+];
+export default catetoryCode;
