+++ src/constants/tagData.js
@@ -1,0 +1,14 @@
+const tagData = [
+  { title: '#광치기 해변🌊', id: 'beach', contentId: 'CNTS_000000000018413' },
+  { title: '#카페코지🍵', id: 'cafe', contentId: 'CNTS_000000000019338' },
+  { title: '#카멜리아힐🌺', id: 'hill', contentId: 'CNTS_000000000001195' },
+  { title: '#휴즐리 제주🍧', id: 'husley', contentId: 'CNTS_300000000015965' },
+  {
+    title: '#제4회 제주비엔날레 : 물과 바람과 별의 길🌟',
+    id: 'vienna',
+    contentId: 'CNTS_300000000013355',
+  },
+  { title: '#파더스가든🐰', id: 'garden', contentId: 'CNTS_200000000014189' },
+  { title: '#비자림🌳', id: 'beejalim', contentId: 'CONT_000000000500270' },
+];
+export default tagData;
