+++ src/pages/AddPlanPage/components/Detail.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+const Detail = ({ onNext, contentId }) => {
+  // 컨텐츠 id로 데이터 가져오기
+  return (
+    <div>
+      Detail
+      <button>일정 생성</button>
+    </div>
+  );
+};
+
+export default Detail;
